[refactor] 중복된 역할을 하는 객체 통합 (LottoPaper <- WinningPaper)
